mrhub/requirements.txt
# --- Core framework ---
Django==4.2.23  # or your chosen LTS

# --- WSGI server for Render (production) ---
gunicorn==23.0.0

# --- Static files (CSS/JS/images) ---
whitenoise==6.9.0

# --- Environment variable loader (local dev, harmless on Render) ---
python-dotenv>=1.0

# --- Database URL parser (Postgres on Render, SQLite locally) ---
dj-database-url>=2.1

# --- HTTP client (for Storacha, APIs, etc.) ---
requests>=2.32

# --- Django admin skin ---
django-grappelli

# --- Django REST Framework ---
djangorestframework==3.16.1

# --- PDF generation ---
reportlab

# --- API schema & docs ---
drf-spectacular

django-cors-headers

psycopg2-binary